refactor(bisheng-core): bisheng-core 拆解方案一初步形态，不接管 webpack 流程。
待讨论的问题：
- getBishengConfig 是否应该进入 bisheng-core；
- 如何整合现有API降低成本，escapeWinPath 这种`src/index`和现在的core中都会用到的代码，依赖关系真的丑。
