+++ packages/bisheng-core/package.json
@@ -1,20 +1,16 @@
 {
-  "name": "bisheng",
+  "name": "bisheng-core",
   "version": "0.28.6",
-  "description": "Transform Markdown(and other static files with transformers) into a SPA website using React.",
-  "bin": {
-    "bisheng": "./bin/bisheng"
-  },
+  "description": "The core of bisheng",
+  "main": "lib/index.js",
   "files": [
-    "lib",
-    "bin",
-    "router.js",
-    "collect.js"
+    "src",
+    "lib"
   ],
   "scripts": {
     "build": "babel src --out-dir lib",
     "clean": "rimraf lib",
-    "prepublish": "npm run build && shx cp src/template.html lib/template.html",
+    "prepublish": "npm run build",
     "postpublish": "npm run clean"
   },
   "babel": {
@@ -69,7 +65,6 @@
     "babel-polyfill": "^6.9.1",
     "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.24.1",
-    "bisheng-core": "^0.28.6",
     "case-sensitive-paths-webpack-plugin": "^2.0.0",
     "chalk": "^1.1.3",
     "commander": "^2.9.0",
