+++ packages/bisheng-core/src/config/getBabelCommonConfig.js
