+++ packages/bisheng-core/src/config/getTSCommonConfig.js
