+++ packages/bisheng-core/src/loaders/common/worker.js
