+++ packages/bisheng-core/src/placeholders/ssr-data.js
