+++ packages/bisheng-core/src/utils/resolve-plugins.js
